Set the maximum number of active database connections.
This number was determined after a series of tests. Increasing it to 16 had similar results but increased slightly quicker than at 8. This is directly related to the number of active threads in MySQL and may not be optimal for configurations where more memory is available and would especially not be optimal if we began using a MySQL cluster.