Added a field in document/read indicating the maximum role for the requesting user for this document.
This field is based on the document's role in all of the campaign and classes to which it is associated as well as the user's document role in their direct association with the document. This may be overridden and increased to the role of "writer" if the requesting user is a supervisor in any of the campaigns to which the document is associated or is privileged in any of the classes to which the document is associated.