Added the rest of the DomainExceptions and their checks to the app.
As a side note, I also reworked how responses are validated and built. This should be a much simpler flow with the only caveat that the response object is just that, an Object. So the type safety is slightly lacking except that if you have a response object you can get the correctly typed response, assuming it wasn't skipped or not displayed.