Fixed a bug in user/read and user/search regarding the placement of the metadata.
This lead to several other transparent but notable changes. The main one is that there has been a reworking of the respond() abstraction in UserRequest and Request. Request now contains only one respond() that adds to an already complete JSONObject the "result"-"success" key-value pair. UserRequest now contains multiple respond() abstractions that all end up calling a single respond() function that refreshes the user's authentication token cookie and then calls the parent's respond(). Also, a new respond() abstraction has been added that allows a user to pass a JSONObject representing the metadata along with the JSONObject that represents the actual data.