Fixed a bug with checking the label values in a prompt's properties.
Moved the check from the Campaign which would blindly do it for all prompt types which wasn't appropriate to only the choice prompts which were the only ones that needed this check. Leaving it Campaign would, for example, prevent a text prompt whose min and max values were the same, but that is acceptable.