Added a preference table item to limit the maximum number of survey responses per page.
The new preference table id is "max_survey_response_page_size", which can be any value indicating the maximum number of survey responses per page. Also, the database scripts have been updated to reflect this change with a default value of "-1". "-1" indicates that there is no limit. As an aside, there was nother bug fixed in the MultiChoiceCustomPrompt class, but it was minor.

Closes #188