Fixed a bug where labels were being incorrectly validated.
The bug pertained to processing the key-label-values for a prompt. The broken way was to ensure that no two labels had the same value. However, that was only the case for choice and custom choice prompts. However, we had a text prompt whose "max" and "min" keys both had a label value of "1" which was perfectly valid, but the validator was complaining.

Closes #208