Changed the word "username" to "user" in the authentication portions of the request.
When performing authentication, the parameter names are "user" and "password"; however, the text of the error message when the user's username is missing is "username". It may be worth looking into in a 3.x release to rename "user" to "username" and giving a different name to the actual "username" parameter used in some requests.