Fixed a bug regarding how timestamp prompt responses were being inserted into the database.
The insertion into the database simply calls getResponse().toString(). However, for timestamp prompt responses the response object would be a Date, and calling toString() on a Date would result in an unexpected formatting for the date. To fix this, I check if the response object is a date and convert it to an ISO 8601-formatted String before insertion.