Fixed a bug where single choice prompt response output was broken.
The columns in the output would be shifted. There were two issues at play here. The first was that the value being treated as the label was sometimes "SKIPPED" or "NOT_DISPLAYED", so when it tried to correlate that "label value" to an actual key-value-label it would fail and null would be output. The second issue was with looking up the value for those "SKIPPED" or "NOT_DISPLAYED" responses. The value in the key-value-label trio for those responses was being set to null which was appropriate, but there was a missing check in the output which, when the value was null, should have appended a ",null" string to the row if the prompt had values.

Closes #212