Fixed a bug regarding the output fields in survey_response/read.
The location's timestamp key in the JSON should have been "location_timestamp" but was instead being set as "timestamp" which was overriding the survey response timestamp which had the same key. This has now been fixed and "time", "timestamp", and "utc_timestamp" refer to the same point in time with the "timezone" needed to correlate the "timestamp". "location_timestamp" and "location_timezone" are the time and timezone for the location object; however, here "location_timestamp" is a long, milliseconds from epoch, value whereas "timestamp" is an ISO8601-formatted timestamp with "time" being the long, milliseconds from epoch, value. For the timezone, "timezone" is the survey response's timezone and "location_timezone" is the location's timezone. We should probably not flatten the location object and should instead have a "location" key that refers to a location JSONObject that contains all of the location information such as the time and timezone.