Added the username of the requesting user to the audit log for all successful  request.
The username for user-password requests was always being audited, but it was being placed in the audit_param table. Now, we are adding the username, key 'user', to the audit_extra table for every successful user request. This is dependant on the User object in the request being populated which must have happened for a successful user request.

Closes #211