Updated how the system preferences were read.
Before, we had an entry in the preference table in the database that pointed to a preference file generated by ant. This is unnecessary because we built and placed the file ourselves, and it is part of our WAR file. Now, the preference cache simply reads where the root of our system is, adds the location to the preference file, and goes from there.