Near complete class/read admin visibility.
Everything appears to work unless the list of users is too long. I tried to use group_concat to group the users and their respective roles in the class/read query to leave all of the aggregation in the database and lower the data transfer overhead. The problem is that group_concat has a hard limit. Instead, many forums suggest that one should consume all of them as different rows and then aggregate the data in Java; however, that prevents us from using LIMIT to limit the results and seems slow overall. I will investigate further but may end up doing what the forums suggets.