Added and tested admin visibility on class/read.
An admin will now see all information about all classes whereas a regular user may or may not see information about a class depending on their role in the class. If a non-admin requests information about a non-existant class or a class that they don't have visibility to view, that class' information (or lack thereof) will be omitted from the result. For example, if a user requests about class A and class B of which they can only view class A, the result will be a JSON array of exactly on JSON object that is the information about class A.