Refactored document/read.
Document read cannot be made such that admins can see all documents. It is constructed, unlike all of the other APIs, with inclusive parameters rather than exclusive. In other words, a user needs to tell the server which groups of documents they should return either through those that are directly linked to the user or those that are passively linked to the user through campaigns and classes. This means that, a user, even if they are an admin, has no way of obtaining the information about all of the documents without sending the list of all campaigns and classes. A requirement in the system is that a document must be shared with at least one campaign or class, so this would solve that problem but at a huge cost.

To remedy this, we need to have document/read passively return all documents with parameters that are exclusive. Therefore, a parameter will limit the list rather than define it. However, this will break backwards compatability, so it may not be feasible until the 3.0 release.