Added BASE64 encoding for image uploading in survey_response/upload.
Survey responses may now be uploaded as either multipart/form-post uploads or as a "regular" application/x-www-form-urlencoded upload. For the former, which is the way it was being done before this, there would be individual attachments to the request, one for each image, where the key was the image's ID and the value was the binary representation of the image. This method is now deprecated. The new way of doing it, the latter of the two, is to add an additional parameter, "images", which is a JSON object where the keys are the image IDs and the values are the BASE64 encoding of the images. Technically, a user could upload a multipart/form-post upload where some images were represented as attachments and others in the "images" parameter; however, this is highly discouraged.