Updated to use Joda time.
The Joda time library does a better job of keeping the time zone list up to date and throws an exception when the given time zone string is unknown. As an aside, this will do many more things that we were previously doing by hand such as creating ISO8601-formatted date strings, so we might want to consider using this library more fully.