Added a "date" column to survey_response/read.
A new column, "urn:ohmage:context:date", has been added to allow the user to only get the date component of the survey response. The main use case for this was to create a way to aggregate information based on the date, i.e. number of survey responses for a single user in a single campaign across all dates.