Added a class role flag and return value to class/read.
class/read now allows for a new parameter, 'class_role', which limits the results to only those where the user is a memeber of the class and has the given role. It now outputs the requesting user's role in the resulting JSON for each class regardless of whether or not the 'class_role' parameter was given. If the user doesn't have a role in a class, the 'role' field in the resulting JSON will be omitted. For the current ACL, this is only possible for admins.

Closes #269