Updated missing parameters on survey/upload.
Before, we were failing the request "by hand" and then throwing the exception as opposed to including the failure information in the exception, which is the way we started doing it a while back. This has been updated to include the failure in the exception as opposed to doing it "by hand" and then throwing the exception.