Fixed the ACLs on document/read.
The ACLs were missing a check for the document's privacy state. The code has shown decent performance; however, it is clunky. The API for document/read needs to be reworked to operate like all of the other APIs, where the base list is all of the documents visible to the user and the parameters limit those documents. Instead, the initial list of documents is based on several parameters which means the list must first be aggregated before its can be widdled down by the ACL, instead of both at the same time. As a result, the code is clunky and potentially inefficient. For now, it has shown decent performance, but this is definitely an API that needs to be rethought.

Closes #295