Added an index to Mobility.
Mobility queries were beginning to slow down, so an index was added to speed them up. Currently, the only additional index is on the time, but when we begin querying on other factors, we might want to consider adding more keys. Preliminary tests show an increase in speed by about a factor of 10.