Added logging to mobility/read.
Mobility read was missing logging statements. Not only do these give a better understanding of what is happening in the logs, but, because each log comes with a millisecond timestamp, it can give us a better idea of what service call is taking the longest when an API begins to slow down.