Timestamp prompt responses may now have an additional format.
The new format is "yyyy-MM-ddTHH:mm:ss", which extends the current set of formats, "yyyy-MM-dd", "yyyy-MM-dd HH:mm:ss", and "yyyy-MM-ddTHH:mm:ssZZ". This was done because the phone was returning this new format, and this was the easiest change for the time being. For 2.12, we shall decide on an official time string and will attempt to standardize on that in the future.