Fixed a bug in audit/read.
The reads were basing themselves on the database timestamp which is incompatible with the milliseconds value it was being given. Now, the start and end dates refer to any request that was in the queue even if the received milliseconds was before the start date but the respond milliseconds were before the end date and visa versa.