Fixed an issue regarding incorrect time zones.
The time zones on location records was being sent as 0. Our previous library was decoding this as the milliseconds offset, and, therefore, setting it to GMT; however, this was not always the case. As a fix for users with this version, we now translate that into what the timezone for the containing object was. Going forward, the clients should no longer have this issue.

Closes #334