Added and tested stream upload and stream read.
There is an issue in the Avro library where integers are not cast to doubles. Therefore, when a schema defines a double value and an integer value is supplied, i.e. "0" instead of "0.0", then the Avro library will claim it is a bug. A fix is supposedly coming for the next release of Avro, 1.7, but I don't have a timeline on when that will occur.