Fixed a potential issue regarding not closing connection streams.
We observered an issue on one of the servers where the number of connections to files was being reached. While the exact cause of this has not yet been fully described, there is a chance that it had to do with the way images and documents were being handled. Their InputStreams may not have been closed. This has been fixed by always attempting to close the file even if an uncaught exception is thrown.