Cleaned up how streams are closed.
Reading the documentation for Streams and Writers in Java, two interesting tid-bits were discovered. First, closing a stream implicitly flushes it. So, I removed the flush in several places where it was unnecessary. Second, closing a stream or writer whose constructor took a stream or writer will, when closed, close that underlying stream or writer. Therefore, I have removed a couple locations where we were closing the parent stream and then closing the underlying stream, which was unnecessary and redundant.

References #344