Added all of the sensor data columns to mobility/read.
The mobility/read shim layer now also outputs the sensor data. We still need to include the classification data, but we haven't yet decided how to store that data. Also, there is a bug where the columns list is being respected by the underlying observer read requests but not by the shim mobility/read layer; instead, it complains very loudly despite the potential that it is a perfectly acceptable scenario.