Added the optional ID to the meta-data.
A user may now specify an ID in the stream's definition's meta-data. When uploading a point with an ID, it will guarantee that no other point for that user and that stream has the same ID. If there is a collision, the new point with the same ID as the old point will be ignored with the assumption that it was already inserted into the database; therefore, it is the responsability of the ID generator to ensure that no ID for the same user in the same stream is created twice.

References #264