Optimized stream data's read API.
MySQL does not understand 'long' as a data type correctly. It translates it into a string type, which causes a serious performance hit. That was converted to a bigint in the observer_stream_data table, and a new column of the same type was created in order to precompute the addition that is performed in the retrieval query. Now, it is only a numeric comparison rather than adding two strings that were first converted to numbers and the result compared with the parameter in the WHERE clause. Finally, an index was placed on that field to further speed up the query. A basic test showed that the same query takes approximately 1/10 the original amount of time after this change.