Added 405's for OMH registry create and update and added observer/read.
OMH has added a new first-class object called "registry". The creation and updating of registry entries is analgous to the creation and updating of observers. However, there are some inconsistencies between the two, so, for now, it was decided that servers may return a HTTP 405 for these calls.

observer/read has been implemented, which gives users the ability to read about all of the observers in the system or limit the search to only those with a certain ID or version.

References #264