Added OMH's registry APIs and migrated from Avro to JSON Schema.
The OMH registry APIs create, read, and update have been implemented. create and update return HTTP 405's because it was decided that we can punt on them for now. The read request and the observer create, read, and update APIs all use the JSON Schema to define probes instead of Avro.

References #350