Refactored OMH read for external parties.
The OMH read API was refactored to be better suited to third party resources. This includes each getting their own PayloadId class and their own Request.

Also, the Mood Map was partially integrated before it was discovered that integration was too costly and complex. This has lead to a simpler implementation, for the time being, where the Mood Map phone application will be ported as a campaign.

References #386