Fixed a bug in mobility/read regarding the 'client' parameter.
There was an issue regarding the 'client' (ohmage) and 'requester' (OMH) parameter when one was omitted, based on the call. Because mobility/read can be called from either side (ohmage or OMH) it was incorrectly altered to only report about the 'requester' and not the 'client'. It will now correctly report about the correct parameter depending on which API was called.