Added RunKeeper to the OMH registry and read.
RunKeeper is now integrated with the OMH component of ohmage. In order to use it, the username of the user on RunKeeper must be the same as the ohmage username. Also, the key exchange must take place and the user's key must be stored in the database's 'omh_authentication' table with the 'domain' "run_keeper" and the 'auth_key' "bearer_" and the username appended on the end. Only two RunKeeper APIs were implemented, "profile" and "fitnessActivities", so the two payload IDs are "omh:run_keeper:profile" and "omh:run_keeper:fitnessActivities".

Closes #381