Fixed a bug regarding WiFi scans with duplicate SSIDs.
For mobility-specific APIs the validation was checking that the WiFi scan did not have duplicate strength values for the same SSID. The data was being uploaded via stream/read, which doesn't check the data this way. Therefore, the read was breaking. This has been fixed by allowing this duplication, which doesn't seem to have a negative impact on the reads.

Closes #410