Added filters for our request Content-Encoding and client/requester parameters.
Two new filters have been added. The first will check if the content encoding on the request exists and is "gzip". If so, it will decode the parameters and store them as an attribute of the request. If not, it will use Tomcat's decoded version of the parameters and stores those as an attribute of the request.

The second will retrieve the "client" parameter, or the "requester" parameter if the "client" parameter doesn't exist, will place it in the NDC, and will save it as an attribute of the request. The requests will then check if this attribute exists when retrieving its value as opposed to decoding it itself as was done before. It should be noted that no requests may have a "client" or "requester" parameter as even having a "requester" parameter in a non-OMH request would causing naming collisions. Technical details aside, this is probably a good idea to prevent confusion, anyway.

Closes #411