Fixed a bug regarding output of non-Jackson JSON objects.
The library we were using was using JSON.org's Java library. That library does not include the appropriate annotations for Jackson to serialize it via its JSON generators. One of the prompt types response value, remote activity, would return the JSON.org objects, which was causing a crash when Jackson attempted to serialize them. This has been fixed, and they are now being appropriately serialized.