Added the Emergency or Provider Contact Thing to HealthVault.
The Emergency or Provider Contact Thing may now be queried through the HealthVault payload ID "omh:health_vault:emergency_or_provider_contact". There was a bug in the library I was using which required me to go and pull from their trunk, so the library has been updated to the trunk snapshot. It passed all of their tests and seems stable. However, if we experience problems with this API, this is probably where we should look first.

Closes #413