Added the columns parameter to the showcase APIs.
All of the APIs that are associated with the showcase now adhere to the column parameter. All of our APIs, campaigns and observers, already did.

The exception is Microsoft's HealthVault. We are using a library that uses annotations, however, we are having a problem using the columns list with Jackson. See: http://stackoverflow.com/questions/12865569/filtering-fields-and-sub-fields-in-jackson

Closes #437