Fixed a few bugs regarding data requests.
Some of the requests were not properly filtering their data based on the start and end date parameters. This has been fixed and now, even though some data may returned from the third-party servers, it is being filtered on our side. The reason that data may be requested from them but we filter it out is because they only allow year, month, and day in their request parameters, but Open mHealth allows for time and time zone components as well.