Updated the FQDN to only be the domain.
The FQDN in the database is now just the domain, e.g. dev.ohmage.org, instead of what it was previously, e.g. https://dev.ohmage.org/. This also means there is no path, e.g. dev.ohmage.org, not dev.ohmage.org/.

When pushing this update, be sure to make the change in the database.

Closes #475