Updated the registration links.
In Outlook, the multi-part messages were causing problems, so it is now one, single HTML document.

There was another formatting issue regarding the "I agree" link (#496). The code is designed to compile several preferences into an email. If there are ever formatting issues in the email, the preference table should be consulted to determine which fields need to be changed.

Closes #502
References #496