Added the default observers.
Now, when a server is first created the default servers will also be created. The update script, however, will not add these observers because it is presumed that they, either, already exist or aren't desired. To add them, go to the db/settings/default_observers.sql file and execute the appropriate lines.

Closes #503