Added the campaign assignment API.
The campaign assignment API allows the requesting user to assign a campaign to another user with a corresponding campaign mask. This mask describes what subset of the surveys the user should see. This means that, if the campaign is visible to a user, only a certain subset of the surveys will be returned in the campaign definition.

This API also allows the requesting user to create a new account. The account will not be an admin and will not be able to create campaigns. It will be ready to use and will not require a password reset.

This API also allows the requesting user to add the user to a class. The class must exist and the requesting user must be privileged in that class. The user being assigned will be restricted in the class. If the user is already associated with the class, their permissions in that class will not change. While the API creates the appropraite campaign/survey mask for the user, the class association makes the campaign visible to the user.

This is an incomplete feature because the campaign/read has not yet been modified to adhere to the mask.

References #493