Added the campaign mask when reading a campaign's definition.
This is done by building the Campaign object, then building the CampaignMask objects, then assigning the CampaignMask objects to the campaign. A better way would be to have them associated, but not necessarily contained in the actual Campaign object. This way, a memcached-like system could cache the campaign definition and the masks independantly, so when a different user asks for the same campaign, the cached campaign can be retrieved and the new masks, if any, could be retrieved and applied independantly.

References #493