Open mHealth's read and registry are now fixed.
The registry was returning results for surveys and prompts that were not being requested. This has now been limited to either return only the survey or prompt requested. For example, "omh:ohmage:campaign:urn:campaign:test:survey_id:mySurvey" will only return the definition for the survey "mySurvey" in the campaign "urn:campaign:test" and not for any of its prompts. Likewise, "omh:ohmage:campaign:urn:campaign:test:survey_id:mySurvey:prompt_id:myPrompt" will only return the definition for the prompt "myPrompt" in the survey "mySurvey" in the campaign "urn:campaign:test".

The read call has also been updated to conform to the registry. This is simply a delayed pull from the showcase branch, which is where all of the Open mHealth APIs were first developed.

Closes #550
Closes #551