Changed the default build target and created a new one.
There is a new build target called "dist-ssl", which is analogous to "dist-no_ssl". The target "dist" will now build the "dist-no_ssl" variant rather than the "dist-ssl" variant. The default target (used when ant is not given any targets) is still "dist-all", which cleans, tests, and builds the SSL and non-SSL versions as well as the documentation.

Closes #222