Created a flow for giving pseudo-anonymized information.
When a user uploads data, they may provide an "opt_in" parameter which will allow their invalid data to be pseudo-anonymized and placed in a special location to which the creator of the observer may have access to it. The creator of the observer may then make a call to stream/invalid/read to get the data.

References #500