Fixed a concurrency bug with class/update.
When deleting a user from a class, if two identical queries were made at near the same time, there was a chance that the second would return a HTTP 500 status code. While the server was never in an inconsistant state, that particular request appeared to have an issue. Now, the code will detect this and skip the user.