Fixed the image issue by creating a background process.
There now exists a background process that will create the image variants. This will allow uploads to still be very fast. In the event that someone asks for a copy of the image that has not yet been processed, it will be processed at the time it is requested and served.

Closes #594
References #451