Added a new status code when an invalid password was given but a valid authentication token was given.
When attempting to change their own password or another user's password, the requesting user must supply their own username and password for security reasons. If the password they provide is incorrect, they will be returned a "0200" error code. Now, if they also happened to supply an authentication token, the token is validated, and, if valid, a new status code, but the same error message, will be returned. That new status code is "0203". This is because legacy systems and those with limited development resources are unable/unwilling to properly handle the "0200" error code.

Closes #604