Fixed a bug when setting up users.
If a user was being added to a class via user/setup to which they were already associated, it was returning an error. Now, it should first check if they are already associated with the class and, if not, it will attempt to reset them by removing them from the class and then immediately readding them.