Renamed Communities to Ohmlets.
Communities are now referred to as ohmlets, and there is a single string to skin the name. Chaning this value will only change the error message text and not the API endpoint or internal representation.

There was also a slight change in account creation. If a user supplies a username and password that already exists, they will be given a success message. This is assumed to only happen when a user attempts to create an account that they previously tried to create due to some error.

References #629
References #632