Updated authorization and survey upload.
All APIs should now be working and in an alpha state. They will be moved to a beta state once the unit tests are written and to a RC state when they have been end-to-end tested. That said, they should not change in syntax or functionality from this point.

Survey responses are now properly being uploaded and saved in the DB. The survey responses can be read but their corresponding media cannot. This is the next step.

Logging and has been updated to work with Log4J2.

Authentication tokens have been removed in favor of authorization tokens. An authentication token is now a specialized authorization token where the scope is global and can only be given to the requesting user; in other words, a third-party can never acquire an authentication token. The OAuth2 authorization flow still needs to be completed. References #485

Media prompts may still give their corresponding duration/size values, however they are not being checked. References #654