Added the location back to the meta-data.
The meta-data may now also contain a location. The location consists of spacial (latitude, longitude, and accuracy) and temporal (timestamp) components. The timestamp of the location should not replace the timestamp of the meta-data. The meta-data's timestamp referes to when the correspding data was created, whereas the location's timestamp refers to when the location reading was made. The two can be compared to determine how long ago the location data was gathered to further decrease its accuracy.

Closes #667