Added the initial version of the Open mHealth APIs.
The current state of the APIs implements the registry and data reading. The authentication was already part of the system. The authorization will be a much larger task. The parts that are still missing are the parameter compliance (e.g. start/end dates, column lists, etc.) and paging. Paging is another large task that will be system-wide.

References #658