Added paging to all list queries.
Any API that returns a list of results is now being paged. This includes returning three headers, "Count", "Previous", and "Next". "Count" is the number of elements being returned in the response. "Previous" and "Next" are fully-qualified URLs that can be used to get the previous and next sets of data, respectively.

Closes #658