Added a testing library and removed user-name.
A testing suite of JUnit and EasyMock has been added to run unit tests. Currently, only the user object and the user servlet are being tested.

The user-name has been removed. An automatically-generated ID has been added. The user-name was for users to have an alternate identity in ohmlets. Now, instead of having a global pseudonym, a user will be (future work) able to have per-ohmlet pseudonyms.

Closes #672