Reworked how data was shared and read.
Data is now being properly shared when read. When a user reads data, it will send them all of the data that is visible for that stream or survey. A user may join and leave ohmlets at any time and the data will be shared and unshared, respectively.

There are two modifications that still need to be made. First, a parameter to the read APIs should be given so that a user may indicate what subset of users they wish to read about. Second, the user may want to specify a pseudonym to use within the ohmlet to prevent cross-ohmlet data aggregation as well as date ranges to ensure that only data within a certain time-frame is shared.

Closes #674