Added application information to streams.
Streams may now list some application information to help onboard a user with an application that corresponds to a particular stream. The stream definition has been updated to have a field "apps" whose value is an array of AppInformation objects. Every AppInformation object must specify a "platform" (e.g. "android", "ios", "windows", etc.) and an "app_uri" which can be used to direct the user to install the application for the respective platform. The "platform" value is not validated to allow for unknown platforms. Optionally, the AppInformation may also include a "authorization_uri" that can be used to direct a user to authorize ohmage to read the user's data.

Closes #679