Updated how lists of elements were deserialized.
Due to type-erasure, Jackson could not properly deserialize parameters like List<MyType>. Instead, it was being deserialized as List<Map>, which would end up throwing a ClassCastException when any part of the code tried to iterate over the elements. This has been "fixed" (circumvented) by declaring the parameter as MyType[]. This prevents the type-erasure and allows Jackson to properly deserialize it.