Added user and ohmlet invitations.
A person in an ohmlet that has the right to invite people to the ohmlet may now create ohmlet invitations. By providing an email address, users will either be automatically added to an ohmlet if they already have an account with that email address or they will be sent an email to join ohmage and that specific ohmlet.

Closes #675
Closes #676