Updated the EasyPost page.
The EasyPost page has several new APIs including the ohmlet APIs. It also memorizes the authorization token between page loads, but not once the window is closed. The output window now follows the user as they scroll down the page.

Also, there was a slight formatting change in the Authorize HTML page and the OhmletServlet Java source.

References #690